+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add doc-tests for function `decompile_to_file`.
 * Add `missing_docs` lint and improve documentation.
 * Avoid generating documentation for private items.
-* Improve CI.
+* Improve CI effectiveness and performance.
 
 ## [0.5.0] - 2025-03-03