+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Improve documentation.
 
+### Fixed
+
+* Fix typo in documentation.
+
 ## [0.5.2] - 2025-03-19
 
 ### Changed