+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Add some filesystem integration tests.
+
+### Changed
+
+* Update dependencies.
+
 ## [0.5.4] - 2025-03-21
 
 ### Fixed