+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* Refactor integration test directory structure.
 * Update dependencies.
 
 ## [0.5.4] - 2025-03-21