+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.5] - 2025-03-27
 
 ### Added
 
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-* Use `impl AsRef<Path>` as `filepath` type in `decompile_to_file` to simplify API.
+* Use `impl AsRef<Path>` as `filepath` type in `decompile_to_file` to simplify the public API.
 * Refactor integration test directory structure.
 * Update dependencies.
 
@@ -172,7 +172,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * First release to be published to [crates.io](https://crates.io/).
 
-[unreleased]: https://github.com/0xdea/haruspex/compare/v0.5.4...HEAD
+[unreleased]: https://github.com/0xdea/haruspex/compare/v0.5.5...HEAD
+
+[0.5.5]: https://github.com/0xdea/haruspex/compare/v0.5.4...v0.5.5
 
 [0.5.4]: https://github.com/0xdea/haruspex/compare/v0.5.3...v0.5.4